requirements.txt
pytest~=7.3.1
requests~=2.31.0
lxml~=4.9.2
bs4~=0.0.1
beautifulsoup4~=4.12.2
allure-pytest